Api basic logic has added

+++ DayProgressApi/Models/Progress.cs
@@ -3,7 +3,9 @@
     public class Progress
     {
         public bool IsWorkingTime { get; set; } = false;
-        public int Percentage { get; set; }
+
+        public double Percentage { get; set; }
+
         public string? Description { get; set; }
     }
 }
